+++ apps/main-app/ts/navigation/AppStackNavigator.tsx
@@ -16,8 +16,8 @@ import { cgnLinkingOptions } from "../features/bonus/cgn/navigation/navigator";
 import { fciLinkingOptions } from "../features/fci/navigation/FciStackNavigator";
 import { idPayLinkingOptions } from "../features/idpay/common/navigation/linking";
 import { IngressScreen } from "../features/ingress/screens/IngressScreen";
+import { itwLinkingOptions } from "../features/itwallet/navigation/linking";
 import { ITW_ROUTES } from "../features/itwallet/navigation/routes";
-import { itwLinkingOptions } from "../features/itwallet/navigation/useItwLinkingOptions";
 import { ITW_CREDENTIAL_OFFER_LINKING_PREFIXES } from "../features/itwallet/offer/utils";
 import { storeLinkingUrl } from "../features/linking/actions";
 import { trackIOOpenedFromUniversalAppLink } from "../features/linking/analytics";
